+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [6682](https://github.com/vegaprotocol/vega/issues/6682) - Add indexes to orders by reference query
 - [6668](https://github.com/vegaprotocol/vega/issues/6668) - Add indexes to trades by buyer/seller
 - [6628](https://github.com/vegaprotocol/vega/issues/6628) - Improve node health check in the wallet
+- [6711](https://github.com/vegaprotocol/vega/issues/6711) - `Anti-whale ersatz` validators reward stake scores
 
 ### 🐛 Fixes
 - [6701](https://github.com/vegaprotocol/vega/issues/6701) - Fix `GraphQL` `API` not returning `x-vega-*` headers